refactor(http): split into alias, client, redirect, route, and server files
## What

- Split `net/http/http.go` (and its test file) into five focused files by concern: `alias.go` (net/http type/const/var aliases), `client.go` (`NewClient`, `ClosingRoundTripper`), `redirect.go` (`IgnoreRedirect`, `SameOrigin`, `SameOriginRedirect`, `IsCrossOriginRedirect`), `route.go` (`ParseServiceMethod`, `Pattern`), and `server.go` (`NewServer`). `http.go` keeps `NewRequestWithContext`, `NewServeMux`, `NewChainedHandlers`, `MaxBytesHandler`, `NewResponseController`, `NewTelemetryHandler`, `StatusText`, and `ParseTime`.
- This is a pure move: no declaration, signature, doc comment, or test assertion changed, only file boundaries and per-file import statements.
- Public API surface, package name, and import path are unchanged.

## Why

- `http.go` had grown to mix unrelated concerns (raw stdlib aliases, client construction, redirect/origin policy, route naming, and server construction) in one file, making it harder to navigate and to scope review of future changes.
- Splitting by concern keeps future diffs (for example, touching only server tuning or only redirect policy) contained to a smaller, single-purpose file.

## Testing

- `package=net/http make lint` - passed
- `package=net/http make specs` - passed (102 tests, race + coverage)
- `go build ./...` - passed
- Independent agent verification (order-independent declaration-set diff of the old `http.go`/`http_test.go` against the concatenated new files) confirmed byte-identical content aside from file boundaries and import reordering, with all 16 original test functions preserved.
