+++ net/http/http_test.go
@@ -5,13 +5,9 @@ import (
 	"testing"
 
 	"github.com/alexfalkowski/go-service/v2/bytes"
-	"github.com/alexfalkowski/go-service/v2/config/options"
 	"github.com/alexfalkowski/go-service/v2/crypto/tls"
-	"github.com/alexfalkowski/go-service/v2/internal/test"
 	"github.com/alexfalkowski/go-service/v2/io"
 	"github.com/alexfalkowski/go-service/v2/net/http"
-	"github.com/alexfalkowski/go-service/v2/net/url"
-	"github.com/alexfalkowski/go-service/v2/strings"
 	"github.com/alexfalkowski/go-service/v2/time"
 	"github.com/stretchr/testify/require"
 )
@@ -33,22 +29,6 @@ func TestMaxBytesHandler(t *testing.T) {
 	handler.ServeHTTP(res, req)
 
 	require.Equal(t, http.StatusOK, res.Code)
-}
-
-func TestNewServerRejectsNegativeTimeoutOption(t *testing.T) {
-	t.Parallel()
-
-	handler := http.HandlerFunc(func(http.ResponseWriter, *http.Request) {})
-
-	for _, key := range []string{"read_timeout", "write_timeout", "idle_timeout", "read_header_timeout"} {
-		t.Run(key, func(t *testing.T) {
-			t.Parallel()
-
-			require.Panics(t, func() {
-				http.NewServer(options.Map{key: "-1s"}, handler)
-			})
-		})
-	}
 }
 
 func TestProtocolsReturnsSupportedHTTPVersions(t *testing.T) {
@@ -94,267 +74,4 @@ func TestTransportReturnsConfiguredRoundTripper(t *testing.T) {
 	require.True(t, transport.Protocols.HTTP1())
 	require.True(t, transport.Protocols.HTTP2())
 	require.True(t, transport.Protocols.UnencryptedHTTP2())
-}
-
-func TestNewServerSetsProtocols(t *testing.T) {
-	t.Parallel()
-
-	handler := http.HandlerFunc(func(http.ResponseWriter, *http.Request) {})
-
-	server := http.NewServer(options.Map{}, handler)
-
-	require.Nil(t, server.HTTP2)
-	require.NotNil(t, server.Protocols)
-	require.True(t, server.Protocols.HTTP1())
-	require.True(t, server.Protocols.HTTP2())
-	require.True(t, server.Protocols.UnencryptedHTTP2())
-}
-
-func TestNewServerSetsHTTP2Options(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name                       string
-		options                    options.Map
-		maxConcurrentStreams       int
-		maxReceiveBufferConnection int
-		maxReceiveBufferStream     int
-	}{
-		{
-			name:                 "max concurrent streams",
-			options:              options.Map{"http2_max_concurrent_streams": "7"},
-			maxConcurrentStreams: 7,
-		},
-		{
-			name:                       "max receive buffer per connection",
-			options:                    options.Map{"http2_max_receive_buffer_per_connection": "3MB"},
-			maxReceiveBufferConnection: int(3 * bytes.MB.Bytes()),
-		},
-		{
-			name:                   "max receive buffer per stream",
-			options:                options.Map{"http2_max_receive_buffer_per_stream": "4MB"},
-			maxReceiveBufferStream: int(4 * bytes.MB.Bytes()),
-		},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			t.Parallel()
-
-			server := http.NewServer(test.options, http.HandlerFunc(func(http.ResponseWriter, *http.Request) {}))
-
-			require.NotNil(t, server.HTTP2)
-			require.Equal(t, test.maxConcurrentStreams, server.HTTP2.MaxConcurrentStreams)
-			require.Equal(t, test.maxReceiveBufferConnection, server.HTTP2.MaxReceiveBufferPerConnection)
-			require.Equal(t, test.maxReceiveBufferStream, server.HTTP2.MaxReceiveBufferPerStream)
-		})
-	}
-}
-
-func TestNewServerRejectsOverflowingHTTP2Options(t *testing.T) {
-	t.Parallel()
-
-	for key, value := range map[string]string{
-		"http2_max_concurrent_streams":            "4294967296",
-		"http2_max_receive_buffer_per_connection": "3GB",
-		"http2_max_receive_buffer_per_stream":     "3GB",
-	} {
-		t.Run(key, func(t *testing.T) {
-			t.Parallel()
-
-			require.Panics(t, func() {
-				http.NewServer(options.Map{key: value}, http.HandlerFunc(func(http.ResponseWriter, *http.Request) {}))
-			})
-		})
-	}
-}
-
-func TestSameOriginRedirectMatchesSchemeHostAndPort(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		want error
-		name string
-		next string
-	}{
-		{name: "same origin", next: "https://example.com/next", want: nil},
-		{name: "same origin host case", next: "https://EXAMPLE.com/next", want: nil},
-		{name: "same origin default port", next: "https://example.com:443/next", want: nil},
-		{name: "different host", next: "https://other.example.com/next", want: http.ErrUseLastResponse},
-		{name: "different scheme", next: "http://example.com/next", want: http.ErrUseLastResponse},
-		{name: "different port", next: "https://example.com:444/next", want: http.ErrUseLastResponse},
-	}
-
-	prev, err := http.NewRequestWithContext(t.Context(), http.MethodGet, "https://example.com/start", http.NoBody)
-	require.NoError(t, err)
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-
-			next, err := http.NewRequestWithContext(t.Context(), http.MethodGet, tt.next, http.NoBody)
-			require.NoError(t, err)
-			err = http.SameOriginRedirect(next, []*http.Request{prev})
-			if tt.want == nil {
-				require.NoError(t, err)
-				return
-			}
-
-			require.ErrorIs(t, err, tt.want)
-		})
-	}
-}
-
-func TestSameOriginRedirectStopsAfterTenRedirects(t *testing.T) {
-	t.Parallel()
-
-	via := make([]*http.Request, 10)
-	for i := range via {
-		request, err := http.NewRequestWithContext(t.Context(), http.MethodGet, "https://example.com/redirect", http.NoBody)
-		require.NoError(t, err)
-
-		via[i] = request
-	}
-
-	next, err := http.NewRequestWithContext(t.Context(), http.MethodGet, "https://example.com/redirect", http.NoBody)
-	require.NoError(t, err)
-
-	require.NoError(t, http.SameOriginRedirect(next, via[:9]))
-
-	err = http.SameOriginRedirect(next, via)
-	require.Error(t, err)
-	require.NotErrorIs(t, err, http.ErrUseLastResponse)
-
-	crossOrigin, err := http.NewRequestWithContext(t.Context(), http.MethodGet, "https://other.example.com/redirect", http.NoBody)
-	require.NoError(t, err)
-	require.ErrorIs(t, http.SameOriginRedirect(crossOrigin, via), http.ErrUseLastResponse)
-}
-
-func TestSameOriginComparesSchemeHostAndPort(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name string
-		prev string
-		next string
-		want bool
-	}{
-		{name: "same origin", prev: "https://example.com/start", next: "https://example.com/next", want: true},
-		{name: "same origin host case", prev: "https://example.com/start", next: "https://EXAMPLE.com/next", want: true},
-		{name: "same origin https default port", prev: "https://example.com/start", next: "https://example.com:443/next", want: true},
-		{name: "same origin http default port", prev: "http://example.com/start", next: "http://example.com:80/next", want: true},
-		{name: "different host", prev: "https://example.com/start", next: "https://other.example.com/next", want: false},
-		{name: "different scheme", prev: "https://example.com/start", next: "http://example.com/next", want: false},
-		{name: "different port", prev: "https://example.com/start", next: "https://example.com:444/next", want: false},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-
-			prev, err := url.Parse(tt.prev)
-			require.NoError(t, err)
-
-			next, err := url.Parse(tt.next)
-			require.NoError(t, err)
-
-			require.Equal(t, tt.want, http.SameOrigin(prev, next))
-		})
-	}
-
-	prev, err := url.Parse("https://example.com/start")
-	require.NoError(t, err)
-
-	next, err := url.Parse("https://example.com/next")
-	require.NoError(t, err)
-
-	require.False(t, http.SameOrigin(nil, next))
-	require.False(t, http.SameOrigin(prev, nil))
-}
-
-func TestIsCrossOriginRedirect(t *testing.T) {
-	t.Parallel()
-
-	prev, err := http.NewRequestWithContext(t.Context(), http.MethodGet, "https://example.com/start", http.NoBody)
-	require.NoError(t, err)
-
-	same, err := http.NewRequestWithContext(t.Context(), http.MethodGet, "https://example.com/next", http.NoBody)
-	require.NoError(t, err)
-	same.Response = &http.Response{Request: prev}
-
-	different, err := http.NewRequestWithContext(t.Context(), http.MethodGet, "https://other.example.com/next", http.NoBody)
-	require.NoError(t, err)
-	different.Response = &http.Response{Request: prev}
-
-	require.False(t, http.IsCrossOriginRedirect(nil))
-	require.False(t, http.IsCrossOriginRedirect(prev))
-	require.False(t, http.IsCrossOriginRedirect(same))
-	require.True(t, http.IsCrossOriginRedirect(different))
-}
-
-func TestClosingRoundTripperClosesBodyWhenRequested(t *testing.T) {
-	t.Parallel()
-
-	rt := http.ClosingRoundTripper(func(*http.Request) (*http.Response, error, bool) {
-		return nil, io.ErrUnexpectedEOF, true
-	})
-	body := &test.TrackedBody{Reader: strings.NewReader("body")}
-	req, err := http.NewRequestWithContext(t.Context(), http.MethodPost, "http://example.com", body)
-	require.NoError(t, err)
-
-	res, err := rt.RoundTrip(req)
-	require.Nil(t, res)
-	require.ErrorIs(t, err, io.ErrUnexpectedEOF)
-	require.True(t, body.Closed)
-}
-
-func TestClosingRoundTripperLeavesDelegatedBodyOpen(t *testing.T) {
-	t.Parallel()
-
-	rt := http.ClosingRoundTripper(func(*http.Request) (*http.Response, error, bool) {
-		return &http.Response{StatusCode: http.StatusOK, Body: http.NoBody, Header: http.Header{}}, nil, false
-	})
-	body := &test.TrackedBody{Reader: strings.NewReader("body")}
-	req, err := http.NewRequestWithContext(t.Context(), http.MethodPost, "http://example.com", body)
-	require.NoError(t, err)
-
-	res, err := rt.RoundTrip(req)
-	require.NoError(t, err)
-	require.NoError(t, res.Body.Close())
-	require.False(t, body.Closed)
-}
-
-func TestIgnoreRedirect(t *testing.T) {
-	t.Parallel()
-
-	err := http.IgnoreRedirect(nil, nil)
-	require.ErrorIs(t, err, http.ErrUseLastResponse)
-}
-
-func TestParseServiceMethod(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name    string
-		url     string
-		method  string
-		service string
-		action  string
-	}{
-		{name: "service route", method: http.MethodGet, url: "/test/hello", service: "test", action: "hello"},
-		{name: "deep service route", method: http.MethodPost, url: "/test/users/123", service: "test", action: "users/123"},
-		{name: "root", method: http.MethodGet, url: "/", service: "root", action: "get"},
-		{name: "single segment", method: http.MethodPost, url: "/health", service: "health", action: "post"},
-	}
-
-	for _, test := range tests {
-		t.Run(test.name, func(t *testing.T) {
-			t.Parallel()
-
-			req := httptest.NewRequestWithContext(t.Context(), test.method, test.url, http.NoBody)
-			service, action := http.ParseServiceMethod(req)
-			require.Equal(t, test.service, service)
-			require.Equal(t, test.action, action)
-		})
-	}
 }
